% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_integrity.R
\name{check_data_integrity}
\alias{check_data_integrity}
\title{Check data integrity}
\usage{
check_data_integrity(x)
}
\arguments{
\item{x}{dataframe with at least colums "lat", "long" and "primary_fur_color"}
}
\value{
Original dataframe if all tests are good. Otherwise stops.
}
\description{
Check data integrity
}
\examples{
# A working example
my_data_example <- data.frame(
  lat = c(40.77, 40.78),
  long = c(-73.95, -73.96),
  primary_fur_color = c("grey", "black")
)
check_data_integrity(my_data_example)
}
